Extracts tests for AbstractApi::get() into separate test file
