+++ tests/Unit/Api/AbstractApiTest.php
@@ -140,42 +140,6 @@ class AbstractApiTest extends TestCase
     /**
      * @covers \Redmine\Api\AbstractApi
      * @test
-     * @dataProvider getJsonDecodingFromGetMethodData
-     */
-    public function testJsonDecodingFromGetMethod($response, $decode, $expected)
-    {
-        $client = $this->createMock(Client::class);
-        $client->method('getLastResponseBody')->willReturn($response);
-        $client->method('getLastResponseContentType')->willReturn('application/json');
-
-        $api = new class ($client) extends AbstractApi {};
-
-        $method = new ReflectionMethod($api, 'get');
-        $method->setAccessible(true);
-
-        // Perform the tests
-        if (is_bool($decode)) {
-            $this->assertSame($expected, $method->invoke($api, 'path', $decode));
-        } else {
-            $this->assertSame($expected, $method->invoke($api, 'path'));
-        }
-    }
-
-    public static function getJsonDecodingFromGetMethodData(): array
-    {
-        return [
-            'test decode by default' => ['{"foo_bar": 12345}', null, ['foo_bar' => 12345]],
-            'test decode by default, JSON decode: false' => ['{"foo_bar": 12345}', false, '{"foo_bar": 12345}'],
-            'test decode by default, JSON decode: true' => ['{"foo_bar": 12345}', true, ['foo_bar' => 12345]],
-            'Empty body, JSON decode: false' => ['', false, false],
-            'Empty body, JSON decode: true' => ['', true, false],
-            'test invalid JSON' => ['{"foo_bar":', true, 'Error decoding body as JSON: Syntax error'],
-        ];
-    }
-
-    /**
-     * @covers \Redmine\Api\AbstractApi
-     * @test
      * @dataProvider getXmlDecodingFromGetMethodData
      */
     public function testXmlDecodingFromRequestMethods($methodName, $response, $decode, $expected)
@@ -190,12 +154,7 @@ class AbstractApiTest extends TestCase
         $method->setAccessible(true);
 
         // Perform the tests
-        if ('get' === $methodName) {
-            $return = $method->invoke($api, 'path', $decode);
-
-            $this->assertInstanceOf(SimpleXMLElement::class, $return);
-            $this->assertXmlStringEqualsXmlString($expected, $return->asXML());
-        } elseif ('delete' === $methodName) {
+        if ('delete' === $methodName) {
             $return = $method->invoke($api, 'path');
 
             $this->assertSame($expected, $return);
@@ -210,9 +169,6 @@ class AbstractApiTest extends TestCase
     public static function getXmlDecodingFromGetMethodData(): array
     {
         return [
-            ['get', '<?xml version="1.0"?><issue/>', null, '<?xml version="1.0"?><issue/>'], // test decode by default
-            ['get', '<?xml version="1.0"?><issue/>', true, '<?xml version="1.0"?><issue/>'],
-            ['get', '<?xml version="1.0"?><issue/>', false, '<?xml version="1.0"?><issue/>'], // test that xml decoding will be always happen
             ['post', '<?xml version="1.0"?><issue/>', null, '<?xml version="1.0"?><issue/>'],
             ['put', '<?xml version="1.0"?><issue/>', null, '<?xml version="1.0"?><issue/>'],
             ['delete', '<?xml version="1.0"?><issue/>', null, '<?xml version="1.0"?><issue/>'],
